DM9000: Get version 1.31 of DM900 from another tree
Update DM9000 to a more recent one

Also added a call to allow setting the DM9000 mac address
at startup from u-boot's environmnet. This allows support
for boards that have a DM9000 without an eeprom.

Without this call, a kernel loaded from NAND will not find
a valid MAC and will fail to load the driver.